+++ app/src/main/java/com/example/movie/myFragments/MainFragment.kt
@@ -17,11 +17,11 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.bumptech.glide.Glide
-import com.example.movie.DetailActivity
 import com.example.movie.R
 import com.example.movie.adapter.MoviesAdapter
 import com.example.movie.model.Movie
 import com.example.movie.model.Singleton
+import com.example.movie.view.DetailActivity
 import com.example.movie.view_model.MovieListViewModel
 import com.example.movie.view_model.ViewModelProviderFactory
